requirements.txt
PyOpenGl
PyOpenGL_accelerate
pillow